log/slog: add example showing how to discard logs
Many other logging libraries ship with a discard handler
(github.com/segmentio/events.Discard, log15.DiscardHandler), and there
is some discussion about how it's not trivial to set this up with
log/slog.

There is a proposal to add a DiscardHandler or similar to log/slog,
but in the meantime, an example can help demonstrate the best way for
users to pass around a Logger that does not log anything.

Updates #62005.

Change-Id: I42650f4ea51e145b55d27823059f1c6d29824366