runtime: treat handleAsyncEvent as a system goroutine
This is a wasm only routine that is currently marked as a user goroutine
which can lead to incorrect runtime stacks and cause issues with the
deadlock detector, as it appears to be a goroutine that is leaking.

It is only treated as a system goroutine sometimes, when not running
user finalizers.

For #64620

Change-Id: I3df0b7a4becc7e5e5cd8fe85ffc6ed8d95850d5a
GitHub-Last-Rev: 27f07f0b47db5988cb7ccdfc3d014cff6f841767
GitHub-Pull-Request: golang/go#64630