cmd/go/internal/toolchain: simplify flag parsing and stop at the first non-flag
The first non-flag argument to 'go run' or 'go install' is a package
path to be installed or run. For 'go run' in particular, any arguments
that follow the package path are to be passed to the user program.

In CL 546635 I changed toolchain.goInstallVersion to look for (and
use) flags given in addition to the package path. However, I neglected
to stop the parsing after the package path. In fixing mistake that, I
also realized that the flag parsing in toolchain.goInstallVersion
duplicates a lot of the existing logic in cmdflag.ParseOne. In the
interest of having fewer different implementations to keep in sync,
let's switch to always using cmdflag.Parse.

This has the side effect of making 'go install' pickier about the
flags passed to it, but arguably that's a good thing anyway: it
removes the ambiguity from commands like
'go install -unknownflag -tags -modcacherw pkg@version', without
the need to establish special-case flag parsing rules beyond
the syntax that the flag package normally accepts.

Fixes #64738.
Updates #64282.
Updates #57001.

Change-Id: I6d3ca41bb85c79c959a053caf059d2eb8bba563b