runtime: fix trace EvGoStop Gosched reason to match function
Currently the stop reason for runtime.Gosched is labeled
"runtime.GoSched" which doesn't actually match the function name. Fix
the label to match the function name.

This change regenerates all the tests except go122-gc-stress.test
because it turns out that isn't generated. A follow-up CL will fix that,
but for now, just replace GoSched with Gosched in the test.

Also, it turns out the cmd/trace/v2 tests are way too strict about
network unblock events, and 3 usually pop up. AFAICT this looks
plausible to me, so just lift the restriction on "up to 2" events.

Change-Id: Id7350132be19119c743c259f2f5250903bf41a04