runtime: track gc roots in the heap profiler
This is very rough prototype code to demonstrate enhancing the heap
profile with garbage collection root metadata for in-use samples. Heap
profiles currently only show the allocation site, but provide no
indication of what may be keeping objects recorded in the profile alive.
Users must either reason through the code themselves, or else resort to
much more expensive (and sensitive) diagnostics via collecting core
dumps. We wanted to see what could be done to enhance the existing
always-on diagnostics.

Our approach works by introducing a "meta" object which holds garbage
collection root metadata, opaque to the garbage collector. Starting from
a root, this metadata is initialized and then propagated via the work
buffer alongside objects to be marked. Each time an object is marked,
and more objects are added to the work buffer (e.g. we are marking an
array of pointers) the metadata is carried along with the new objects to
mark.  When we mark an object for the first time, we look for a heap
profile "special" associated with the object. If we find one, we add the
root metadata to a list of roots associated with the sample. Then, at
serialization time, the root metadata objects are resolved to textual
labels. The label contents depend on the kind of root. For goroutine
stack roots, for example, the label is a call stack leading to the call
frame from which the object was reachable. The root labeling is not
deterministic.  Whichever root reaches an object first wins for that
object. That is, this is not computing a dominator (if one exists) for
each tracked object.

This is absolutely hacky WIP proof-of-concept research code written
during a team R&D week. We have observed non-negligible overhead even
with heap profiling disabled due solely to the cost of tracking the root
metadata. This is merely meant to start a discussion about whether this
kind of enhancement is worth pursuing. We think it may have promise if it
can be done cheaply.

DO NOT REVIEW
DO NOT SUBMIT

Co-Authored-By: Daniel Schwartz-Narbonne <daniel.schwartznarbonne@datadoghq.com>
Change-Id: I4749570da1002b4fb4c04febfa1402f93c2e4a72