cmd/compile/internal/ssa: fix location lists bug
Before this change there was a bug which can occur when an SSA'd
variable has multiple different values (slots) corresponding to the same
piece of variable. Before this change, the locations of each type of
a piece would be concatenated, resulting in nonsensical location
information with more pieces than there ought to be for the data type.

This patch rectifies the situation by associating the slots
corresponding to the same piece and computing the location of that piece
by coalescing the locations of the corresponding slots.

Fixes #61700

Change-Id: I21dfd5408a5a4c6f3a5586511fa041822e09095f
GitHub-Last-Rev: ab3dfbb82cc464c1b0b652d516d53b79d7030690
GitHub-Pull-Request: golang/go#64944