doc(contribution): count of contributors updated
**Description:**
This pull request introduces an update to the contributor count for the Go programming language. The count has been increased from 'hundreds' to 'thousands', reflecting the significant growth in the number of contributors to the Go project.

The updated count accurately represents the current state of the Go community, highlighting the increasing interest and participation in contributing to the Go project. This update is crucial for maintaining transparency and fostering a sense of community among Go developers.

Change-Id: I3ef4b99d56200c5e1ace0adffec66cb39a639756
GitHub-Last-Rev: 291c9e65aec1b045f1a755ed131df2359b01a4b2
GitHub-Pull-Request: golang/go#65022