runtime/race: add test parallel rangefunc exit flag test
This tests the assertion that a rangefunc iterator running
in parallel can trigger the race detector if any of the
parallel goroutines attempts an early exit.  It also
verifies that if everything else is carefully written,
that it does NOT trigger the race detector if all the
parts run time completion.

Change-Id: Ie91b4817de5f6fca98a153efd552db3f78b85e76