go/types, types2: bail early if we want a type but don't have one
If we do know whether we need a type or not, make use of the
information when we know that we don't have a type and bail out.

Fixes the issue at hand and also improves some other error messages
which now report that we don't have a type instead of reporting a cycle.

Also: add an extra parameter to writeObject to control whether a type
is written or not. For shorter error messages and to avoid accessing
a type when it may not be ready yet.

Fixes #65344.

Change-Id: I11182efd452c485d89e6c09ead8a647ea05d7318