net/http: set a client timeout in TestServerWriteTimeout
This breaks an unbounded client-side retry loop if the server's
timeout happens to fire during its final read of the TLS handshake.

The retry loop was observed on wasm platforms at CL 557437.

I was also able to reproduce chains of dozens of retries on my
linux/amd64 workstation by adjusting some timeouts and adding a couple
of sleeps, as in this patch:
https://gist.github.com/bcmills/d0a0a57e5f64eebc24e8211d8ea502b3

However, on linux/amd64 on my workstation the test always eventually
breaks out of the retry loop due to timing jitter.

Separately, we may want to consider some kind of client-side hook to
allow the user to customize the retry behavior to avoid this kind of
unbounded retry loop or implement a backoff policy.

Fixes #65410.
Updates #65178.

Change-Id: I0e43c039615fe815f0a4ba99a8813c48b1fdc7e6